+++ src/ipc/ipc4/notification.c
@@ -12,7 +12,9 @@
 #include <ipc4/notification.h>
 #include <sof/ipc/notification_pool.h>
 
-#include <rtos/symbol.h>
+#ifdef CONFIG_SOF_USERSPACE_LL
+#include <zephyr/internal/syscall_handler.h>
+#endif
 
 static uint32_t notification_mask = 0xFFFFFFFF;
 
@@ -37,8 +39,13 @@ static bool is_notif_filtered_out(uint32_t event_type)
 	return (notification_mask & BIT(notif_idx)) == 0;
 }
 
-static bool send_resource_notif(uint32_t resource_id, uint32_t event_type, uint32_t resource_type,
-				void *data, uint32_t data_size)
+#ifdef CONFIG_SOF_USERSPACE_LL
+bool z_impl_send_resource_notif(uint32_t resource_id, uint32_t event_type,
+				uint32_t resource_type, void *data, uint32_t data_size)
+#else
+bool send_resource_notif(uint32_t resource_id, uint32_t event_type,
+			 uint32_t resource_type, void *data, uint32_t data_size)
+#endif
 {
 	struct ipc_msg *msg;
 
@@ -80,49 +87,34 @@ static bool send_resource_notif(uint32_t resource_id, uint32_t event_type, uint3
 	return true;
 }
 
-static enum sof_ipc4_resource_event_type dir_to_xrun_event(enum sof_ipc_stream_direction dir)
-{
-	return (dir == SOF_IPC_STREAM_PLAYBACK) ? SOF_IPC4_GATEWAY_UNDERRUN_DETECTED :
-						  SOF_IPC4_GATEWAY_OVERRUN_DETECTED;
-}
-
 void ipc4_update_notification_mask(uint32_t ntfy_mask, uint32_t enabled_mask)
 {
 	notification_mask &= enabled_mask | (~ntfy_mask);
 	notification_mask |= enabled_mask & ntfy_mask;
 }
 
-bool send_copier_gateway_xrun_notif_msg(uint32_t pipeline_id, enum sof_ipc_stream_direction dir)
+#ifdef CONFIG_SOF_USERSPACE_LL
+static inline bool z_vrfy_send_resource_notif(uint32_t resource_id, uint32_t event_type,
+					      uint32_t resource_type, void *data,
+					      uint32_t data_size)
 {
-	return send_resource_notif(pipeline_id, dir_to_xrun_event(dir), SOF_IPC4_PIPELINE, NULL,
-				   0);
+	/* Validate event_type is a known resource event */
+	K_OOPS(K_SYSCALL_VERIFY(event_type < SOF_IPC4_INVALID_RESOURCE_EVENT_TYPE));
+
+	/* Validate resource_type is a known resource type */
+	K_OOPS(K_SYSCALL_VERIFY(resource_type < SOF_IPC4_INVALID_RESOURCE_TYPE));
+
+	/* data and data_size must be consistent */
+	K_OOPS(K_SYSCALL_VERIFY((!data && !data_size) || (data && data_size)));
+
+	/* Payload must fit in the event_data union and the IPC message */
+	K_OOPS(K_SYSCALL_VERIFY(data_size <= sizeof(union ipc4_resource_event_data)));
+	K_OOPS(K_SYSCALL_VERIFY(data_size <= SOF_IPC_MSG_MAX_SIZE));
+
+	if (data && data_size)
+		K_OOPS(K_SYSCALL_MEMORY_READ(data, data_size));
+
+	return z_impl_send_resource_notif(resource_id, event_type, resource_type, data, data_size);
 }
-
-bool send_gateway_xrun_notif_msg(uint32_t resource_id, enum sof_ipc_stream_direction dir)
-{
-	return send_resource_notif(resource_id, dir_to_xrun_event(dir), SOF_IPC4_GATEWAY, NULL, 0);
-}
-
-void send_mixer_underrun_notif_msg(uint32_t resource_id, uint32_t eos_flag, uint32_t data_mixed,
-				   uint32_t expected_data_mixed)
-{
-	struct ipc4_mixer_underrun_event_data mixer_underrun_data;
-
-	mixer_underrun_data.eos_flag = eos_flag;
-	mixer_underrun_data.data_mixed = data_mixed;
-	mixer_underrun_data.expected_data_mixed = expected_data_mixed;
-
-	send_resource_notif(resource_id, SOF_IPC4_MIXER_UNDERRUN_DETECTED, SOF_IPC4_PIPELINE,
-			    &mixer_underrun_data, sizeof(mixer_underrun_data));
-}
-EXPORT_SYMBOL(send_mixer_underrun_notif_msg);
-
-void send_process_data_error_notif_msg(uint32_t resource_id, uint32_t error_code)
-{
-	struct ipc4_process_data_error_event_data error_data;
-
-	error_data.error_code = error_code;
-
-	send_resource_notif(resource_id, SOF_IPC4_PROCESS_DATA_ERROR, SOF_IPC4_MODULE_INSTANCE,
-			    &error_data, sizeof(error_data));
-}
+#include <zephyr/syscalls/send_resource_notif_mrsh.c>
+#endif
